requirements.txt
anthropic>=0.39.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
PyJWT>=2.8.0
